docs: Update Shorebird TOS for 2024 (#65)
* docs: Update Shorebird TOS for 2024

This replaces our previous 2023 TOS for all customers.

Our 2023 TOS were written by me (a non-lawyer), cribbed from what
expo.dev was using for their Terms at the time.  This set is
written by an actual lawyer, with the goals of:
1. Be as much like every other Saas.
2. Make it easy for companies to adopt Shorebird.

Terms are currently scheduled to go into affect on March 4, 2024.
All customers will be required to make an affirmative accept to continue using the services.